+++ coco-manager/coco-manager-service/src/main/java/com/coco/service/ItemCatService.java
@@ -1,5 +1,6 @@
 package com.coco.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.coco.common.pojo.EUTreeNode;
@@ -7,9 +8,9 @@ import com.coco.common.pojo.TaotaoResult;
 
 public interface ItemCatService {
 
-	List<EUTreeNode> getCatList(long parentId);
-	TaotaoResult insertItemCat(long parentId, String name);
-	TaotaoResult deleteItemCat(long id);
-	TaotaoResult updateItemCat(long id,String name);
-	boolean judgeDeleteItemCat(long id);
+	List<EUTreeNode> getCatList(BigDecimal parentId);
+	TaotaoResult insertItemCat(BigDecimal parentId, String name);
+	TaotaoResult deleteItemCat(BigDecimal id);
+	TaotaoResult updateItemCat(BigDecimal id,String name);
+	boolean judgeDeleteItemCat(BigDecimal id);
 }
